+++ src/code/browser/service/explorerTree/explorerItem.ts
