+++ src/code/platform/notebook/electron/notebookGroup.ts
@@ -7,8 +7,8 @@ import { join, resolve } from "src/base/common/file/path";
 import { URI } from "src/base/common/file/uri";
 import { Iterable } from "src/base/common/util/iterable";
 import { Strings } from "src/base/common/util/string";
-import { ExplorerItem } from "src/code/browser/workbench/actionView/explorer/explorerItem";
-import { IExplorerOpenEvent } from "src/code/browser/workbench/actionView/explorer/explorerTree";
+import { ExplorerItem } from "src/code/browser/service/explorerTree/explorerItem";
+import { IExplorerOpenEvent } from "src/code/browser/service/folderTree/folderTree";
 import { Notebook } from "src/code/platform/notebook/browser/notebook";
 import { IFileService } from "src/code/platform/files/common/fileService";
 import { createService } from "src/code/platform/instantiation/common/decorator";
